Move viewer/editor to separate module
